Move familiarstate to types
